Add Ghidra project and C header/source for v2.09
This driver version is much closer to the version actually running in
the stock Android build (v2.08). The main focus in this Ghidra project
was understanding the dmesg output, understanding how buffer flipping
worked, and confirming that my data path reimplementation[1] was still
correct.

Specifically, I was not concerned with the custom ioctl API, as that
will not be used on mainline, and as the userspace side (the hwcomposer)
is available as C source. So this version does not fully supersede the
previous effort.

[1]: https://gitlab.com/smaeul/ebc-dev